Break out classes into their own files
Split `HOGFeatureProducer` and `FeatureProducer` into their own files.
Also clean up the unused functions around them. These will have to be
reimplemented in the base class as needed.
